Restructure namespace from azure.monitor to microsoft.azureMonitor
Move Azure Monitor OpenTelemetry modules from azure/monitor/opentelemetry/
to microsoft/azureMonitor/ namespace. Update configure module and tests.
