+++ pyproject.toml
@@ -34,7 +34,7 @@ target-version = ["py310"]
 
 [tool.pylint.main]
 py-version = "3.10"
-ignore-paths = ["src/azure"]
+ignore-paths = ["src/microsoft/azureMonitor"]
 
 [tool.pylint.format]
 max-line-length = 120
