+++ src/microsoft/azureMonitor/__init__.py
@@ -8,7 +8,7 @@
 # is discoverable when this package is installed in editable/development mode.
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)
 
-from azure.monitor.opentelemetry._configure import configure_azure_monitor
+from microsoft.azureMonitor._configure import configure_azure_monitor
 
 from ._version import VERSION
