+++ src/microsoft/azureMonitor/_browser_sdk_loader/_config.py
