+++ src/microsoft/azureMonitor/_browser_sdk_loader/django_middleware.py
