+++ src/microsoft/azureMonitor/_browser_sdk_loader/snippet_injector.py
