+++ src/microsoft/azureMonitor/_constants.py
