+++ src/microsoft/azureMonitor/_diagnostics/__init__.py
