+++ src/microsoft/azureMonitor/_types.py
