+++ src/microsoft/azureMonitor/_version.py
