+++ src/microsoft/azureMonitor/py.typed
